[TASK] Remove not needed TemplateService code
The TemplateService class provides its own wrap()
function which is exactly the same as in cObj.
The only instances where the wrap() method is used
is in the MenuContentObjects - however, they can
also access the wrap() function via cObj.

The MenuContentObjects are now using the cObj
function so the duplicate code in TemplateService
can be deprecated.

Resolves: #55567
Releases: 6.3
Change-Id: Iaa7055473f814acc73429f31772d1946c66004f9
Reviewed-on: http://review.typo3.org/27244
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>