[BUGFIX] Catch all resource exceptions when parsing RTE links
The RteHtmlParser checks for valid links by trying to retrieve
a file or folder object. The FAL API will throw exceptions
if anything goes wrong with the given path.

Catch all these exceptions to ignore the invalid link
and act the same for invalid paths as for not found files/ folders

Releases: 6.2, 6.3
Resolves: #60316
Change-Id: I0752678eb8ed2b596c3a88536d957ec35c172d7b
Reviewed-on: http://review.typo3.org/31619
Reviewed-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Tested-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>